artists graph is done
