+++ client/src/pages/Statstest.js
@@ -1,0 +1,14 @@
+import React from "react";
+import MainNav from "../components/MainNav";
+import ArtistsTotal from "../components/ArtistsTotal";
+
+function Home() {
+  return (
+    <div>
+      <MainNav />
+      <ArtistsTotal />
+    </div>
+  );
+}
+
+export default Home;
